pkg/k8s: refactor CRD handling
This patch refactors CRD handling.

It introduces a new package (crdutils) that includes a CRD type, meant
to hold all the information for registering a CRD.

Registration code is moved into crdutils, and the caller just needs to
specify the CRD objects to register.
